# perf: use column_values in read buffer