Closes #857

## Goal

Maximise utilisation of the Read Buffer and unlock the hawtness!

![](https://thumbs.gfycat.com/PleasantBlondIcterinewarbler-max-1mb.gif)

Really though, just connect the optimised Read Buffer `tag_values` implementation up to the InfluxRPC frontend 😄 

## Benchmarks

Here are the relevant improvements to the benchmarks I added to IOx in #1192. (Anything that didn't change by 5% is not shown).

```
group                                                                                master                                 read_buffer
-----                                                                                ------                                 -----------
Data in open chunk of mutable buffer, and one chunk of read buffer/tag0/no_pred      2.19    106.9±4.25ms        ? ?/sec    1.00     48.9±2.15ms        ? ?/sec
Data in open chunk of mutable buffer, and one chunk of read buffer/tag0/with_pred    5.13     33.3±1.24ms        ? ?/sec    1.00      6.5±0.27ms        ? ?/sec
Data in open chunk of mutable buffer, and one chunk of read buffer/tag1/no_pred      2.12    135.8±4.64ms        ? ?/sec    1.00     64.1±1.63ms        ? ?/sec
Data in open chunk of mutable buffer, and one chunk of read buffer/tag1/with_pred    5.22     33.4±1.40ms        ? ?/sec    1.00      6.4±0.27ms        ? ?/sec
Data in open chunk of mutable buffer, and one chunk of read buffer/tag2/no_pred      1.83    223.1±8.25ms        ? ?/sec    1.00    122.2±5.50ms        ? ?/sec
Data in open chunk of mutable buffer, and one chunk of read buffer/tag2/with_pred    3.12     19.4±0.75ms        ? ?/sec    1.00      6.2±0.33ms        ? ?/sec
Data in two read buffer chunks/tag0/no_pred                                          46.99   114.0±4.50ms        ? ?/sec    1.00      2.4±0.08ms        ? ?/sec
Data in two read buffer chunks/tag0/with_pred                                        48.87    45.1±1.14ms        ? ?/sec    1.00   922.5±49.26µs        ? ?/sec
Data in two read buffer chunks/tag1/no_pred                                          45.81   148.5±6.43ms        ? ?/sec    1.00      3.2±0.11ms        ? ?/sec
Data in two read buffer chunks/tag1/with_pred                                        58.43    44.9±1.04ms        ? ?/sec    1.00   767.9±33.26µs        ? ?/sec
Data in two read buffer chunks/tag2/no_pred                                          23.95   224.1±5.50ms        ? ?/sec    1.00      9.4±0.26ms        ? ?/sec
Data in two read buffer chunks/tag2/with_pred                                        51.19    27.9±0.81ms        ? ?/sec    1.00   545.2±20.52µs        ? ?/sec
```

### Improvements

Now the frontend uses the Read Buffer's dedicated operation for `tag_values` the performance when querying against read buffer chunks has improved between `20-50x`. 

## Follow on Work

 - [ ] I believe that I have left an optimisation out that will allow us to effectively just clone an RLE dictionary directly when we know all rows will match the query (e.g., when there is no predicate and no null values in the tag key column). Tracked in #1207
 - [ ] The `tag_values` implementation is not utilising the "current results" buffer I added to the Read Buffer implementation (and I think it's in the MUB too). This optimisation allows the operation to skip entire row groups / tables / chunks if the current results already contain all the distinct values in the column. Tracked in #1208